Refactor create account
